+++ package.json
@@ -4,13 +4,18 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "$(npm bin)/webpack && node app.js"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "express": "^4.13.0",
     "jade": "^1.11.0",
-    "lodash": "^3.10.0"
+    "jquery": "3.2.1",
+    "lodash": "^3.10.0",
+    "three": "0.89.0"
+  },
+  "devDependencies": {
+    "webpack": "3.10.0"
   }
 }
