+++ webpack.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  entry: './src/cuspid.js',
+  output: {
+    filename: './static/bundle.js'
+  }
+};
